optimize code

+++ app/src/main/java/com/example/stores/MainActivity.kt
@@ -1,10 +1,14 @@
 package com.example.stores
 
+import android.content.Intent
+import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.widget.Toast
 import androidx.recyclerview.widget.GridLayoutManager
 import com.example.stores.databinding.ActivityMainBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import java.util.concurrent.LinkedBlockingQueue
 
 class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
@@ -84,13 +88,68 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
     }
 
     override fun onDeleteStore(storeEntity: StoreEntity) {
-        storeEntity.isFavorite = !storeEntity.isFavorite
-        val queue = LinkedBlockingQueue<StoreEntity>()
-        Thread {
-            StoreApplication.database.storeDao().deleteStore(storeEntity)
-            queue.add(storeEntity)
-        }.start()
-        mAdapter.delete(queue.take())
+        val items = resources.getStringArray(R.array.array_options_item)
+        MaterialAlertDialogBuilder(
+            this
+        )
+            .setTitle(R.string.dialog_options_title)
+            .setItems(items, {dialogInterface, i ->
+                when(i){
+                    0 -> confirmDelete(storeEntity)
+                    1 -> dial(storeEntity.phone)
+                    2 -> goToWebsite(storeEntity.website)
+                }
+            })
+            .show()
+
+    }
+
+    private fun dial(phone: String){
+        var callIntent = Intent().apply{
+            action = Intent.ACTION_DIAL
+            data = Uri.parse("tel: $phone")
+        }
+        startIntent(callIntent)
+
+    }
+
+    private fun goToWebsite(website: String){
+        if(website.isEmpty()){
+            Toast.makeText(this, R.string.main_error_not_website, Toast.LENGTH_LONG).show()
+        }else{
+            var websiteIntent = Intent().apply{
+                action = Intent.ACTION_VIEW
+                data = Uri.parse(website)
+
+            }
+            startIntent(websiteIntent)
+        }
+
+    }
+
+    private fun startIntent(intent: Intent){
+        if(intent.resolveActivity(packageManager) != null){
+            startActivity(intent)
+        }else{
+            Toast.makeText(this, R.string.main_error_no_resolve, Toast.LENGTH_LONG).show()
+        }
+    }
+
+    private fun confirmDelete(storeEntity: StoreEntity){
+        MaterialAlertDialogBuilder(
+            this
+        )
+            .setTitle(R.string.dialog_delete_title)
+            .setPositiveButton(R.string.dialog_delete_confirm, {dialogInterface, i ->
+                val queue = LinkedBlockingQueue<StoreEntity>()
+                Thread {
+                    StoreApplication.database.storeDao().deleteStore(storeEntity)
+                    queue.add(storeEntity)
+                }.start()
+                mAdapter.delete(queue.take())
+            })
+            .setNegativeButton(R.string.dialog_delete_cancel, null)
+            .show()
     }
 
     override fun hideFab(isVisible: Boolean) {
